Allow unschedulable sessions
Sessions that are entered before the scheduler are run are now
immutable with regards to their timeslot. They will not get scheduled
elsewhere.  This enable you to add sessions that have to be in a
certain place at a certain time. The scheduler pays no-attention to them,
so you only want to place them into Timeslots where schedulable is
false.  For example you may place a "Presenters Luncheon, rm 123"
during the  "Lunch" timeslot.

Also added a rake task `app:clear_schedule` This enables the generated
schedule to be cleared so an alternate schedule can be generated.