Allow rooms to be schedulable
Unschedulable rooms can be used for events on the schedule but that you
don't want to be automatically scheduled by the annealer.  For example,
we could add the cafeteria as a room, but we don't want to schedule
sessions there.  We could "hard code" an event (i.e. BOF meetup) to happen
in the cafeteria though.