Make multiple blank users rather than just one.
Presently if you make a bunch of sessions with a blank name, they are
all assigned to the same presenter.  This poses a problem if you later
want to change the presenters name.  It would change the name for all
unassigned sessions. This makes one presenter for each unassigned
session. This allows us to set different names at a later time.